Added /Models folder, and moved all references to this new hiearchy
