+++ app/Auth/MagicAuthentication.php
@@ -3,7 +3,7 @@
 namespace App\Auth;
 
 use Illuminate\Http\Request;
-use App\User;
+use App\Models\User;
 
 class MagicAuthentication
 {
