+++ app/Mail/MagicLoginRequested.php
@@ -5,7 +5,7 @@ namespace App\Mail;
 use Illuminate\Bus\Queueable;
 use Illuminate\Mail\Mailable;
 use Illuminate\Queue\SerializesModels;
-use App\User;
+use App\Models\User;
 
 class MagicLoginRequested extends Mailable
 {
