+++ app/Policies/ClientPolicy.php
@@ -3,7 +3,7 @@
 namespace App\Policies;
 
 use App\Client;
-use App\User;
+use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
 class ClientPolicy
