+++ app/Policies/PaymentPolicy.php
@@ -3,7 +3,7 @@
 namespace App\Policies;
 
 use App\Payment;
-use App\User;
+use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
 class PaymentPolicy
