+++ app/Policies/TransactionPolicy.php
@@ -3,7 +3,7 @@
 namespace App\Policies;
 
 use App\Transaction;
-use App\User;
+use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
 class TransactionPolicy
