+++ resources/assets/js/components/passport/AuthorizedClients.vue
@@ -65,18 +65,21 @@
                 tokens: []
             };
         },
+
         /**
          * Prepare the component (Vue 1.x).
          */
         ready() {
             this.prepareComponent();
         },
+
         /**
          * Prepare the component (Vue 2.x).
          */
         mounted() {
             this.prepareComponent();
         },
+
         methods: {
             /**
              * Prepare the component (Vue 2.x).
@@ -84,6 +87,7 @@
             prepareComponent() {
                 this.getTokens();
             },
+
             /**
              * Get all of the authorized tokens for the user.
              */
@@ -93,6 +97,7 @@
                         this.tokens = response.data;
                     });
             },
+
             /**
              * Revoke the given token.
              */
