+++ tests/TestCase.php
@@ -10,7 +10,7 @@ abstract class TestCase extends BaseTestCase
 
     public function signIn($user = null)
     {
-        $user = $user ?: create('App\User');
+        $user = $user ?: create('App\Models\User');
         $this->actingAs($user);
 
         return $this;
